introduce a ChunkRenderer
Speeds up world rendering a bit since we don't use a different program
(and vertex buffer) object for each drawn chunk.
