+++ client/src/world/mod.rs
@@ -1,5 +1,8 @@
 /// Only intergration of children
+
+mod chunk_renderer;
 mod chunk_view;
 mod world_view;
 
+pub use self::chunk_renderer::*;
 pub use self::world_view::WorldView;
